parse entity file in web worker
