+++ packages/map/package.json
@@ -22,6 +22,7 @@
     "@noita-explorer/fastlz": "*",
     "@noita-explorer/model": "*",
     "@noita-explorer/model-noita": "*",
+    "@noita-explorer/scrapers": "*",
     "@noita-explorer/tools": "*"
   },
   "devDependencies": {
