+++ packages/map/src/main.ts
@@ -1,15 +1,12 @@
-export { renderChunk } from './chunks/render-chunk.ts';
-
-export type { PixelCalculator } from './interfaces/pixel-calculator.ts';
 export type {
   ChunkRenderableEntity,
   ChunkRenderableEntitySprite,
-} from './chunks/chunk-renderable-entity.ts';
+} from './interfaces/chunk-renderable-entity.ts';
 
-// new exports
 export type { ChunkBorders } from './interfaces/chunk-borders.ts';
 export type { NoitaBackgroundTheme } from './interfaces/noita-background-theme.ts';
 export { mapConstants, CAVE_LIMIT_Y } from './map-constants.ts';
 export { renderBiomeTile } from './render/biomes/render-biome-tile.ts';
 export { renderTerrainTile } from './render/terrain/render-terrain-tile.ts';
 export { renderBackgroundTile } from './render/background/render-background-tile.ts';
+export { parseEntityFile } from './parse/entity/parse-entity-file.ts';
